requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.9.0
pytest==8.3.2
httpx==0.27.2
email-validator==2.1.0
